Unify variable types across project; use common type for storing numer of pages

+++ sdk/xtdk/xtfw.h
@@ -117,7 +117,7 @@ typedef struct _LOADER_MEMORY_MAPPING
     LIST_ENTRY ListEntry;
     PVOID VirtualAddress;
     PVOID PhysicalAddress;
-    UINT NumberOfPages;
+    ULONGLONG NumberOfPages;
     LOADER_MEMORY_TYPE MemoryType;
 } LOADER_MEMORY_MAPPING, *PLOADER_MEMORY_MAPPING;
